Fix build generation
