+++ next.config.js
@@ -24,5 +24,5 @@ const nextConfig = {
   },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
